Cast the amounts to be summed to be longs.

+++ PrintLogApi/Profiles/FilamentProfile.cs
@@ -14,15 +14,15 @@ namespace PrintLogApi.Profiles
                 .ForMember(dest => dest.LoadedInPrinter, src => src.MapFrom(src => src.PrinterFilaments.Where(pf => !pf.UnloadedDateTime.HasValue).Select(p => p.Printer).FirstOrDefault()))
                 .ForMember(dest => dest.FilamentRemaining, src => src.MapFrom(src => (src.InitialNominalWeightMg ?? 0)
                                                                                     - src.PrintFilaments.Sum(p => p.AmountMg.HasValue && p.AmountMg > 0 ?
-                                                                                                                    p.AmountMg :
+                                                                                                                    (long) p.AmountMg :
                                                                                                                     p.EstimatedAmountMg.HasValue && p.EstimatedAmountMg > 0 ?
-                                                                                                                    p.EstimatedAmountMg : 0)
+                                                                                                                    (long) p.EstimatedAmountMg : (long)0)
                                                                                     + src.FilamentAdjustments.Sum(adj => adj.AmountMg)))
                 .ForMember(dest => dest.FilamentLengthRemainingInM, src => src.MapFrom(src => src.DiameterMm > 0 && src.MaterialDensityGramPerCubicCm > 0 ? (((src.InitialNominalWeightMg ?? 0)
                                                                                     - src.PrintFilaments.Sum(p => p.AmountMg.HasValue && p.AmountMg > 0 ?
-                                                                                                                    p.AmountMg :
+                                                                                                                    (long)p.AmountMg :
                                                                                                                     p.EstimatedAmountMg.HasValue && p.EstimatedAmountMg > 0 ?
-                                                                                                                    p.EstimatedAmountMg : 0)
+                                                                                                                    (long)p.EstimatedAmountMg : (long)0)
                                                                                     + src.FilamentAdjustments.Sum(adj => adj.AmountMg)) ?? 0) / (250 * Math.PI * src.MaterialDensityGramPerCubicCm * src.DiameterMm * src.DiameterMm) : 0))
                 .ForMember(dest => dest.FilamentVolumeRemainingInMl, src => src.MapFrom(src => ((src.InitialNominalVolumeMl ?? 0)
                                                                                     - src.PrintFilaments.Sum(p => p.VolumeMl.HasValue && p.VolumeMl > 0 ?
